requirements.txt
Django==1.5.1
django-tastypie==0.9.15
